# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coherence API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/embed": {
      "post": {
        "tags": [
          "embed"
        ],
        "summary": "Embed",
        "operationId": "embed_embed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resonance": {
      "post": {
        "tags": [
          "resonance"
        ],
        "summary": "Resonance",
        "operationId": "resonance_resonance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResonanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResonanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/axes/{pack_id}/activate": {
      "post": {
        "tags": [
          "axes"
        ],
        "summary": "Activate Pack",
        "operationId": "activate_pack_v1_axes__pack_id__activate_post",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/axes/{pack_id}": {
      "get": {
        "tags": [
          "axes"
        ],
        "summary": "Get Pack",
        "operationId": "get_pack_v1_axes__pack_id__get",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/axes/{pack_id}/export": {
      "get": {
        "tags": [
          "axes"
        ],
        "summary": "Export Pack",
        "description": "Export full axis pack vectors as JSON for inline testing.\n\nNote: Large payloads; intended for dev/test only.",
        "operationId": "export_pack_v1_axes__pack_id__export_get",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/axes/{axis_pack_id}": {
      "get": {
        "tags": [
          "axes"
        ],
        "summary": "Get Pack",
        "description": "Get a single axis pack summary by ID.",
        "operationId": "get_pack_axes__axis_pack_id__get",
        "parameters": [
          {
            "name": "axis_pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Axis Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Pack Axes  Axis Pack Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivateResponse": {
        "properties": {
          "active": {
            "additionalProperties": true,
            "type": "object",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "active"
        ],
        "title": "ActivateResponse"
      },
...
      "EmbedRequest": {
        "properties": {
          "texts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Texts",
            "description": "List of input texts"
          },
          "encoder_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encoder Name",
            "description": "HF model id override"
          },
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device",
            "description": "\"cpu\" | \"cuda\" | \"mps\" | \"auto\""
          },
          "normalize_input": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalize Input",
            "description": "Lowercase/strip inputs deterministically"
          }
        },
        "type": "object",
        "required": [
          "texts"
        ],
        "title": "EmbedRequest"
      },
      "EmbedResponse": {
        "properties": {
          "embeddings": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Embeddings"
          },
          "shape": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Shape"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "device": {
            "type": "string",
            "title": "Device"
          }
        },
        "type": "object",
        "required": [
          "embeddings",
          "shape",
          "model_name",
          "device"
        ],
        "title": "EmbedResponse"
      },
...
      "ExportResponse": {
        "properties": {
          "pack_id": {
            "type": "string",
            "title": "Pack Id"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Names"
          },
          "Q": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Q"
          },
          "lambda_": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Lambda"
          },
          "beta": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Beta"
          },
          "weights": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Weights"
          }
        },
        "type": "object",
        "required": [
          "pack_id",
          "names",
          "Q",
          "lambda_",
          "beta",
          "weights"
        ],
        "title": "ExportResponse"
      },
...
      "GetResponse": {
        "properties": {
          "pack_id": {
            "type": "string",
            "title": "Pack Id"
          },
          "dim": {
            "type": "integer",
            "title": "Dim"
          },
          "k": {
            "type": "integer",
            "title": "K"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Names"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "pack_hash": {
            "type": "string",
            "title": "Pack Hash"
          }
        },
        "type": "object",
        "required": [
          "pack_id",
          "dim",
          "k",
          "names",
          "meta",
          "pack_hash"
        ],
        "title": "GetResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResonanceRequest": {
        "properties": {
          "vectors": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vectors",
            "description": "(n,d) vectors; if omitted, texts must be provided"
          },
          "texts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Texts",
            "description": "Texts to auto-embed; used if vectors not provided"
          },
          "axis_pack": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AxisPackModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Id",
            "description": "Optional axis pack id to load from server registry"
          },
          "return_intermediate": {
            "type": "boolean",
            "title": "Return Intermediate",
            "default": false
          },
          "encoder_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encoder Name"
          },
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "normalize_input": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalize Input"
          }
        },
        "type": "object",
        "title": "ResonanceRequest"
      },
      "ResonanceResponse": {
        "properties": {
          "scores": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Scores"
          },
          "coords": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coords"
          },
          "utilities": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utilities"
          }
        },
        "type": "object",
        "required": [
          "scores"
        ],
        "title": "ResonanceResponse"
      },
...
    }
  }
}